# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "TuckZone Canteen API",
        "description": "Backend for the school canteen pre-order app.\n\nAuthentication: obtain tokens from /api/auth/login or\n/api/auth/otp/login, then send `Authorization: Bearer <accessToken>`.\nWhen a call returns 401, exchange the refresh token at\n/api/auth/refresh rather than forcing the user to sign in again.\n\nOrdering is idempotent: generate one idempotencyKey when the\ncheckout screen opens and reuse it for every retry of that\nsame order, otherwise a double tap creates two orders.",
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/api/admin/orders/{id}/status": {
            "put": {
                "tags": [
                    "order-controller"
                ],
                "operationId": "adminTransition",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderStatusUpdateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/orders": {
            "get": {
                "tags": [
                    "order-controller"
                ],
                "operationId": "myOrders",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrderResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "order-controller"
                ],
                "operationId": "placeOrder",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PlaceOrderRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/orders/collect": {
            "post": {
                "tags": [
                    "order-controller"
                ],
                "operationId": "collect",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "pickupCode",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/ordering/open": {
            "post": {
                "tags": [
                    "ordering-window-controller"
                ],
                "operationId": "open",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderingWindowRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderingWindowResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/ordering/close": {
            "post": {
                "tags": [
                    "ordering-window-controller"
                ],
                "operationId": "close",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderingWindowRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderingWindowResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "order-controller"
                ],
                "operationId": "getMyOrder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "order-controller"
                ],
                "operationId": "cancelOrder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/ordering-status": {
            "get": {
                "tags": [
                    "ordering-window-controller"
                ],
                "operationId": "status",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrderingWindowResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/orders": {
            "get": {
                "tags": [
                    "order-controller"
                ],
                "operationId": "adminListOrders",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "PLACED",
                                "ACCEPTED",
                                "PREPARING",
                                "PACKED",
                                "OUT_FOR_DELIVERY",
                                "DELIVERED",
                                "REJECTED",
                                "CANCELLED"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrderResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/demand": {
            "get": {
                "tags": [
                    "ordering-window-controller"
                ],
                "operationId": "demand",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DemandRow"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "OrderStatusUpdateRequest": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "PLACED",
                            "ACCEPTED",
                            "PREPARING",
                            "PACKED",
                            "OUT_FOR_DELIVERY",
                            "DELIVERED",
                            "REJECTED",
                            "CANCELLED"
                        ]
                    },
                    "deliveryPersonName": {
                        "type": "string",
                        "maxLength": 120,
                        "minLength": 0
                    }
                },
                "required": [
                    "status"
                ]
            },
            "OrderItemResponse": {
                "type": "object",
                "properties": {
                    "menuItemId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "itemName": {
                        "type": "string"
                    },
                    "unitPrice": {
                        "type": "number"
                    },
                    "quantity": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "lineTotal": {
                        "type": "number"
                    }
                }
            },
            "OrderResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "orderNumber": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "PLACED",
                            "ACCEPTED",
                            "PREPARING",
                            "PACKED",
                            "OUT_FOR_DELIVERY",
                            "DELIVERED",
                            "REJECTED",
                            "CANCELLED"
                        ]
                    },
                    "orderType": {
                        "type": "string",
                        "enum": [
                            "DELIVERY",
                            "TAKEAWAY"
                        ]
                    },
                    "pickupCode": {
                        "type": "string"
                    },
                    "menuDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "slotName": {
                        "type": "string"
                    },
                    "deliveryTime": {
                        "type": "string"
                    },
                    "recipientName": {
                        "type": "string"
                    },
                    "deliveryLocation": {
                        "type": "string"
                    },
                    "deliveryPersonName": {
                        "type": "string"
                    },
                    "paymentMethod": {
                        "type": "string",
                        "enum": [
                            "WALLET",
                            "DIRECT"
                        ]
                    },
                    "paymentStatus": {
                        "type": "string",
                        "enum": [
                            "PENDING",
                            "PAID",
                            "REFUNDED"
                        ]
                    },
                    "totalAmount": {
                        "type": "number"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItemResponse"
                        }
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "OrderLineRequest": {
                "type": "object",
                "properties": {
                    "menuItemId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "quantity": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1
                    }
                },
                "required": [
                    "menuItemId",
                    "quantity"
                ]
            },
            "PlaceOrderRequest": {
                "type": "object",
                "properties": {
                    "slotId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "orderType": {
                        "type": "string",
                        "enum": [
                            "DELIVERY",
                            "TAKEAWAY"
                        ]
                    },
                    "menuDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "beneficiaryStudentProfileId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "deliveryLocation": {
                        "type": "string",
                        "maxLength": 200,
                        "minLength": 0
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderLineRequest"
                        },
                        "minItems": 1
                    },
                    "idempotencyKey": {
                        "type": "string",
                        "maxLength": 80,
                        "minLength": 0
                    }
                },
                "required": [
                    "idempotencyKey",
                    "items",
                    "menuDate",
                    "slotId"
                ]
            },
...
            "OrderingWindowRequest": {
                "type": "object",
                "properties": {
                    "menuDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "slotId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "overrideCutoffTime": {
                        "type": "string"
                    },
                    "reason": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 0
                    }
                },
                "required": [
                    "menuDate",
                    "slotId"
                ]
            },
            "OrderingWindowResponse": {
                "type": "object",
                "properties": {
                    "menuDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "slotId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "slotName": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "OPEN",
                            "CLOSED"
                        ]
                    },
                    "effectiveCutoffTime": {
                        "type": "string"
                    },
                    "acceptingOrders": {
                        "type": "boolean"
                    },
                    "reason": {
                        "type": "string"
                    }
                }
            },
...
            "DemandRow": {
                "type": "object",
                "properties": {
                    "menuItemId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "itemName": {
                        "type": "string"
                    },
                    "orderedQuantity": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalQuantity": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "remainingQuantity": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "shortfall": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}